refactor(checks): rename BaseInteractionSpec to BaseInteraction and update related imports
- Renamed `BaseInteractionSpec` to `BaseInteraction` for clarity and consistency.
- Updated all relevant imports and references throughout the codebase to reflect this change.
- Adjusted type hints in various components to ensure compatibility with the new naming convention.
- Enhanced the overall structure of interaction handling in the checks module.
