+++ libs/giskard-checks/src/giskard/checks/core/__init__.py
@@ -1,13 +1,15 @@
 from .check import Check
+from .interaction import BaseInteraction, Interaction
 from .result import CheckResult, CheckStatus, Metric, ScenarioResult, TestCaseResult
 from .scenario import Scenario
 from .testcase import TestCase
-from .trace import Interaction, Trace
+from .trace import Trace
 
 __all__ = [
     "Scenario",
     "Trace",
     "Interaction",
+    "BaseInteraction",
     "Check",
     "CheckResult",
     "CheckStatus",
