+++ libs/giskard-checks/tests/scenarios/test_testcase.py
@@ -12,7 +12,6 @@ from giskard.checks import (
     CheckResult,
     Equals,
     Interaction,
-    InteractionSpec,
     TestCase,
     Trace,
 )
@@ -391,7 +390,7 @@ class TestTestCaseEdgeCases:
             return f"Processed: {inputs}"
 
         trace = await Trace.from_interactions(
-            InteractionSpec(inputs="test", outputs=output_func)
+            Interaction(inputs="test", outputs=output_func)
         )
         check = Equals(
             expected_value="Processed: test",
